requirements.txt
alembic==1.4.2
Babel==2.8.0
blinker==1.4
Bootstrap-Flask==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
dominate==2.5.1
elasticsearch==7.8.0
email-validator==1.1.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
guess-language-spirit==0.5.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.5
PyJWT==1.7.1
PyMySQL==0.10.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
requests==2.24.0
rq==1.5.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
flask-httpauth==4.1.0
Flask-SocketIO==4.3.1
eventlet==0.26.1
